Move instance into servicecatalog/client
